cli: refactor ensure-log-options
